+++ FreeRTOS/Demo/RISC-V_Galois_P1/bsp/xilinx/axiethernet/xaxiethernet.c
@@ -1735,7 +1735,7 @@ void XAxiEthernet_PhyRead(XAxiEthernet *InstancePtr, u32 PhyAddress,
 	volatile s32 TimeoutLoops;
 
 	printf("XAxiEthernet_PhyRead: BaseAddress: 0x%08x, register: 0x%lx\r\n",
-		InstancePtr->Config.BaseAddress, RegisterNum);
+		(size_t)(InstancePtr->Config.BaseAddress), RegisterNum);
 
 	/*
 	 * Verify that each of the inputs are valid.
